Move deserialization into domain and fix tests.
